Move React stuff out of marker schema logic module.
This was causing all of react to be bundled into the symbolicator-cli script.
