+++ src/components/tooltip/Marker.tsx
@@ -37,8 +37,8 @@ import {
 import { Backtrace } from 'firefox-profiler/components/shared/Backtrace';
 
 import {
-  formatMarkupFromMarkerSchema,
   getSchemaFromMarker,
+  formatFromMarkerSchema,
 } from 'firefox-profiler/profile-logic/marker-schema';
 import { computeScreenshotSize } from 'firefox-profiler/profile-logic/marker-data';
 
@@ -52,6 +52,7 @@ import type {
   PageList,
   MarkerSchemaByName,
   MarkerIndex,
+  MarkerFormatType,
   InnerWindowID,
   Page,
   Pid,
@@ -59,6 +60,7 @@ import type {
   IndexIntoStackTable,
 } from 'firefox-profiler/types';
 
+import type { StringTable } from 'firefox-profiler/utils/string-table';
 import type { ConnectedProps } from 'firefox-profiler/utils/connect';
 import {
   getGCMinorDetails,
@@ -281,7 +283,7 @@ class MarkerTooltipContents extends React.PureComponent<Props> {
           const displayLabel = this.props.showKeys ? key : label || key;
           details.push(
             <TooltipDetail key={schema.name + '-' + key} label={displayLabel}>
-              {formatMarkupFromMarkerSchema(
+              {renderMarkerFieldValue(
                 schema.name,
                 format,
                 value,
@@ -555,6 +557,134 @@ class MarkerTooltipContents extends React.PureComponent<Props> {
         {this._maybeRenderNetworkPhases()}
       </div>
     );
+  }
+}
+
+// This regexp is used to test for URLs and remove their scheme for display.
+const URL_SCHEME_REGEXP = /^http(s?):\/\//;
+
+/**
+ * This function may return structured markup for some types suchs as table,
+ * list, or urls. For other types this falls back to formatFromMarkerSchema
+ * above.
+ */
+export function renderMarkerFieldValue(
+  markerType: string,
+  format: MarkerFormatType,
+  value: any,
+  stringTable: StringTable,
+  threadIdToNameMap?: Map<Tid, string>,
+  processIdToNameMap?: Map<Pid, string>
+): React.ReactElement | string {
+  if (value === undefined || value === null) {
+    console.warn(`Formatting ${value} for ${JSON.stringify(markerType)}`);
+    return '(empty)';
+  }
+  if (format !== 'url' && typeof format !== 'object' && format !== 'list') {
+    return formatFromMarkerSchema(
+      markerType,
+      format,
+      value,
+      stringTable,
+      threadIdToNameMap,
+      processIdToNameMap
+    );
+  }
+  if (typeof format === 'object') {
+    switch (format.type) {
+      case 'table': {
+        const { columns } = format;
+        if (!(value instanceof Array)) {
+          throw new Error('Expected an array for table type');
+        }
+        const hasHeader = columns.some((column) => column.label);
+        return (
+          <table className="marker-table-value">
+            {hasHeader ? (
+              <thead>
+                <tr>
+                  {columns.map((col, i) => (
+                    <th key={i}>{col.label || ''}</th>
+                  ))}
+                </tr>
+              </thead>
+            ) : null}
+            <tbody>
+              {value.map((row, i) => {
+                if (!(row instanceof Array)) {
+                  throw new Error('Expected an array for table row');
+                }
+
+                if (row.length !== columns.length) {
+                  throw new Error(
+                    `Row ${i} length doesn't match column count (row: ${row.length}, cols: ${columns.length})`
+                  );
+                }
+                return (
+                  <tr key={i}>
+                    {row.map((cell, i) => {
+                      return (
+                        <td key={i}>
+                          {renderMarkerFieldValue(
+                            markerType,
+                            columns[i].type || 'string',
+                            cell,
+                            stringTable,
+                            threadIdToNameMap,
+                            processIdToNameMap
+                          )}
+                        </td>
+                      );
+                    })}
+                  </tr>
+                );
+              })}
+            </tbody>
+          </table>
+        );
+      }
+      default:
+        throw new Error(
+          `Unknown format type ${JSON.stringify(format as never)}`
+        );
+    }
+  }
+  switch (format) {
+    case 'list':
+      if (!(value instanceof Array)) {
+        throw new Error('Expected an array for list format');
+      }
+      return (
+        <ul className="marker-list-value">
+          {value.map((_entry, i) => (
+            <li key={i}>
+              {renderMarkerFieldValue(
+                markerType,
+                'string',
+                value[i],
+                stringTable
+              )}
+            </li>
+          ))}
+        </ul>
+      );
+    case 'url': {
+      if (!URL_SCHEME_REGEXP.test(value)) {
+        return value;
+      }
+      return (
+        <a
+          href={value}
+          target="_blank"
+          rel="noreferrer"
+          className="marker-link-value"
+        >
+          {value.replace(URL_SCHEME_REGEXP, '')}
+        </a>
+      );
+    }
+    default:
+      throw new Error(`Unknown format type ${JSON.stringify(format as never)}`);
   }
 }
